Extract InMemoryBaseTrie unchanged in preparation for other trie types
